# scheduler/internal/input/http/gen/types.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package gen

// Defines values for Status.
const (
	Completed Status = "completed"
	Failed    Status = "failed"
	Queued    Status = "queued"
	Running   Status = "running"
)

// Execution defines model for Execution.
type Execution struct {
	FinishedAt *int64  `json:"finishedAt,omitempty"`
	Id         *string `json:"id,omitempty"`
	JobId      *string `json:"jobId,omitempty"`
	StartedAt  *int64  `json:"startedAt,omitempty"`
	Status     *string `json:"status,omitempty"`
	WorkerId   *string `json:"workerId,omitempty"`
}

// Job defines model for Job.
type Job struct {
	CreatedAt int64                  `json:"createdAt"`
	Id        string                 `json:"id"`
	Interval  *string                `json:"interval,omitempty"`
	Once      *string                `json:"once,omitempty"`
	Payload   map[string]interface{} `json:"payload"`
	Status    Status                 `json:"status"`
}

// JobCreate defines model for JobCreate.
type JobCreate struct {
	Interval *string                 `json:"interval,omitempty"`
	Once     *string                 `json:"once,omitempty"`
	Payload  *map[string]interface{} `json:"payload,omitempty"`
}

// Status defines model for Status.
type Status string

// GetJobsParams defines parameters for GetJobs.
type GetJobsParams struct {
	Status *Status `form:"status,omitempty" json:"status,omitempty"`
}

// GetJobsJobIdExecutionsParams defines parameters for GetJobsJobIdExecutions.
type GetJobsJobIdExecutionsParams struct {
	WorkerId *string `form:"worker_id,omitempty" json:"worker_id,omitempty"`
}

// PostJobsJSONRequestBody defines body for PostJobs for application/json ContentType.
type PostJobsJSONRequestBody = JobCreate